split models and add new ones
