+++ src/app/models/movie-genres.model.ts
@@ -1,0 +1,9 @@
+// Interfaces for movie genres
+export interface MovieGenre {
+  id: number;
+  name: string;
+}
+
+export interface MovieGenreList {
+  genres: MovieGenre[];
+}
